clean up fiberref
