+++ core/js/src/main/scala/zio/RuntimeConfigPlatformSpecific.scala
@@ -64,7 +64,7 @@ private[zio] trait RuntimeConfigPlatformSpecific {
         level: LogLevel,
         message: () => String,
         cause: Cause[Any],
-        context: Map[FiberRef.Runtime[_], AnyRef],
+        context: Map[FiberRef[_], AnyRef],
         spans: List[LogSpan],
         annotations: Map[String, String]
       ) => {
