+++ core/jvm/src/main/scala/zio/FiberPlatformSpecific.scala
@@ -43,7 +43,7 @@ private[zio] trait FiberPlatformSpecific {
           }
         }
 
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
 
       def id: FiberId = FiberId.None
 
@@ -80,7 +80,7 @@ private[zio] trait FiberPlatformSpecific {
           }
         }
 
-      def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
+      def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
 
       def id: FiberId = FiberId.None
