+++ core/shared/src/main/scala/zio/Fiber.scala
@@ -147,7 +147,7 @@ sealed abstract class Fiber[+E, +A] { self =>
    * Gets the value of the fiber ref for this fiber, or the initial value of the
    * fiber ref, if the fiber is not storing the ref.
    */
-  def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A]
+  def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A]
 
   /**
    * The identity of the fiber.
@@ -248,8 +248,8 @@ sealed abstract class Fiber[+E, +A] { self =>
       final def await(implicit trace: ZTraceElement): UIO[Exit[E1, B]] =
         self.await.flatMap(_.foreach(f))
 
-      final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]]   = self.children
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] = self.getRef(ref)
+      final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]] = self.children
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A]       = self.getRef(ref)
 
       def id: FiberId = self.id
       final def inheritRefs(implicit trace: ZTraceElement): UIO[Unit] =
@@ -285,7 +285,7 @@ sealed abstract class Fiber[+E, +A] { self =>
 
       final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]] = self.children
 
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] =
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A] =
         for {
           first  <- self.getRef(ref)
           second <- that.getRef(ref)
@@ -460,7 +460,7 @@ sealed abstract class Fiber[+E, +A] { self =>
 
       final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]] = self.children
 
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] =
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A] =
         (self.getRef(ref) zipWith that.getRef(ref))(ref.join(_, _))
 
       final def id: FiberId = self.id <> that.id
@@ -674,7 +674,7 @@ object Fiber extends FiberPlatformSpecific {
         IO.foreachPar(fibers)(_.await.flatMap(IO.done(_))).exit
       final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]] =
         ZIO.foreachPar(Chunk.fromIterable(fibers))(_.children).map(_.flatten)
-      def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] =
+      def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A] =
         UIO.foldLeft(fibers)(ref.initial)((a, fiber) => fiber.getRef(ref).map(ref.join(a, _)))
 
       final def id: FiberId = fibers.foldLeft(FiberId.None: FiberId)(_ <> _.id)
@@ -710,13 +710,13 @@ object Fiber extends FiberPlatformSpecific {
    */
   def done[E, A](exit: => Exit[E, A]): Fiber.Synthetic[E, A] =
     new Fiber.Synthetic[E, A] {
-      final def await(implicit trace: ZTraceElement): UIO[Exit[E, A]]                      = IO.succeedNow(exit)
-      final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]]   = IO.succeedNow(Chunk.empty)
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
-      final def id: FiberId                                                                = FiberId.None
-      final def interruptAs(id: FiberId)(implicit trace: ZTraceElement): UIO[Exit[E, A]]   = IO.succeedNow(exit)
-      final def inheritRefs(implicit trace: ZTraceElement): UIO[Unit]                      = IO.unit
-      final def poll(implicit trace: ZTraceElement): UIO[Option[Exit[E, A]]]               = IO.succeedNow(Some(exit))
+      final def await(implicit trace: ZTraceElement): UIO[Exit[E, A]]                    = IO.succeedNow(exit)
+      final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]] = IO.succeedNow(Chunk.empty)
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A]       = ZIO.succeed(ref.initial)
+      final def id: FiberId                                                              = FiberId.None
+      final def interruptAs(id: FiberId)(implicit trace: ZTraceElement): UIO[Exit[E, A]] = IO.succeedNow(exit)
+      final def inheritRefs(implicit trace: ZTraceElement): UIO[Unit]                    = IO.unit
+      final def poll(implicit trace: ZTraceElement): UIO[Option[Exit[E, A]]]             = IO.succeedNow(Some(exit))
     }
 
   /**
@@ -789,7 +789,7 @@ object Fiber extends FiberPlatformSpecific {
 
       final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]] = ZIO.succeedNow(Chunk.empty)
 
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A] = ZIO.succeed(ref.initial)
 
       final def id: FiberId = FiberId.None
 
@@ -883,7 +883,7 @@ object Fiber extends FiberPlatformSpecific {
     new Fiber.Synthetic[Nothing, Nothing] {
       final def await(implicit trace: ZTraceElement): UIO[Exit[Nothing, Nothing]]                    = ZIO.never
       final def children(implicit trace: ZTraceElement): UIO[Chunk[Fiber.Runtime[_, _]]]             = ZIO.succeedNow(Chunk.empty)
-      final def getRef[A](ref: FiberRef.Runtime[A])(implicit trace: ZTraceElement): UIO[A]           = ZIO.succeed(ref.initial)
+      final def getRef[A](ref: FiberRef[A])(implicit trace: ZTraceElement): UIO[A]                   = ZIO.succeed(ref.initial)
       final def id: FiberId                                                                          = FiberId.None
       final def interruptAs(id: FiberId)(implicit trace: ZTraceElement): UIO[Exit[Nothing, Nothing]] = ZIO.never
       final def inheritRefs(implicit trace: ZTraceElement): UIO[Unit]                                = ZIO.unit
